+++ scripts/run_e2e_tests_test.py
@@ -29,6 +29,7 @@ from scripts import common
 from scripts import flake_checker
 from scripts import install_third_party_libs
 from scripts import run_e2e_tests
+from scripts import scripts_test_utils
 
 import contextlib2
 
@@ -42,7 +43,7 @@ def mock_managed_process(*unused_args, **unused_kwargs):
         Context manager. A context manager that always yields a mock
         process.
     """
-    return contextlib2.nullcontext(enter_result=test_utils.PopenStub())
+    return contextlib2.nullcontext(enter_result=scripts_test_utils.PopenStub())
 
 
 class RunE2ETestsTests(test_utils.GenericTestBase):
@@ -258,7 +259,8 @@ class RunE2ETestsTests(test_utils.GenericTestBase):
     def test_work_with_non_ascii_chars(self):
         def mock_managed_protractor_server(**unused_kwargs): # pylint: disable=unused-argument
             return contextlib2.nullcontext(
-                enter_result=test_utils.PopenStub(stdout='sample\n✓\noutput\n'))
+                enter_result=scripts_test_utils.PopenStub(
+                    stdout='sample\n✓\noutput\n'))
 
         self.exit_stack.enter_context(self.swap_with_checks(
             run_e2e_tests, 'is_oppia_server_already_running', lambda *_: False))
